requirements/dev.txt
-r common.txt
django-debug-toolbar
git+git://github.com/kaleidos/django-sampledatahelper.git
